# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is a user-friendly solution designed to serve as the foundation for more complex APIs.",
        "title": "Go API",
        "contact": {
            "name": "Raul del Aguila",
            "url": "https://github.com/raulaguila",
            "email": "email@email.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/auth": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User authenticated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User authenticated",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.UserOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "User refresh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User refresh",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "boolean",
                        "description": "Expire token",
                        "name": "expire",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.AuthOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "User authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User authentication",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "Credentials model",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.AuthOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                },
                "object": {}
            }
        },
        "github_com_raulaguila_go-api_internal_core_dto.AuthOutput": {
            "type": "object",
            "properties": {
                "accesstoken": {
                    "type": "string"
                },
                "refreshtoken": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.UserOutput"
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.LoginInput": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "expiration": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.RoleOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.UserOutput": {
            "type": "object",
            "properties": {
                "corp_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "new": {
                    "type": "boolean"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.RoleOutput"
                    }
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}